map_variables: move to scan_utils
